pyproject.toml
[project]
name = "thermography"
version = "0.1.0"
description = "Professional thermography website built with Django + Wagtail CMS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.1",
    "django-csp>=4.0",
    "django-storages[google,s3]>=1.14.6",
    "gunicorn>=25.1.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "sib-api-v3-sdk>=7.6.0",
    "wagtail>=7.3",
    "whitenoise>=6.11.0",
]
